Update SSN to 32 bit
For some reason we had set the SSN as 16 bit, id can be above 65535 so we need to store as 32 bit.

+++ BT_Demo/SmartRadio.m
@@ -31,7 +31,7 @@
 @property (strong, nonatomic) CBMutableCharacteristic   *rateCharacteristic;
 @property (strong, nonatomic) CBMutableCharacteristic   *dataTxCharacteristic;
 @property (strong, nonatomic) CBMutableCharacteristic   *dataRxCharacteristic;
-@property (nonatomic) UInt16                            SSN;
+@property (nonatomic) UInt32                            SSN;
 
 @property (nonatomic, assign) id                        delagate;
